backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
google-generativeai==0.8.0
youtube-transcript-api==0.6.2
python-multipart==0.0.9
